+++ compiler/src/org.graalvm.compiler.truffle.common/src/org/graalvm/compiler/truffle/common/TruffleCompilerRuntime.java
@@ -405,26 +405,11 @@ public interface TruffleCompilerRuntime {
     ResolvedJavaType resolveType(MetaAccessProvider metaAccess, String className, boolean required);
 
     /**
-     * Gets the option values for this runtime as a map from option name to option value.
-     */
-    Map<String, Object> getOptions();
-
-    /**
-     * Gets the option values for this runtime in an instance of {@code type}.
+     * Gets the Graal option values for this runtime in an instance of {@code type}.
      *
      * @throws IllegalArgumentException if this runtime does not support {@code type}
      */
-    default <T> T getOptions(Class<T> type) {
-        throw new IllegalArgumentException(getClass().getName() + " can not return option values of type " + type.getName());
-    }
-
-    /**
-     * Convert option values in name/value pairs to an instance of {@code type}.
-     *
-     * @param map input option values as {@link String} names to values
-     * @throws IllegalArgumentException if this runtime does not support {@code type}
-     */
-    default <T> T convertOptions(Class<T> type, Map<String, Object> map) {
+    default <T> T getGraalOptions(Class<T> type) {
         throw new IllegalArgumentException(getClass().getName() + " can not return option values of type " + type.getName());
     }