+++ compiler/src/org.graalvm.compiler.truffle.runtime.hotspot.java/src/org/graalvm/compiler/truffle/runtime/hotspot/java/HotSpotTruffleRuntime.java
@@ -24,17 +24,10 @@
  */
 package org.graalvm.compiler.truffle.runtime.hotspot.java;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import org.graalvm.collections.EconomicMap;
-import org.graalvm.collections.UnmodifiableMapCursor;
+import org.graalvm.compiler.debug.DebugOptions;
 import org.graalvm.compiler.hotspot.CompilerConfigurationFactory;
 import org.graalvm.compiler.hotspot.HotSpotGraalOptionValues;
-import org.graalvm.compiler.options.OptionDescriptors;
-import org.graalvm.compiler.options.OptionKey;
 import org.graalvm.compiler.options.OptionValues;
-import org.graalvm.compiler.options.OptionsParser;
 import org.graalvm.compiler.truffle.common.TruffleCompiler;
 import org.graalvm.compiler.truffle.compiler.hotspot.HotSpotTruffleCompilerImpl;
 import org.graalvm.compiler.truffle.compiler.hotspot.HotSpotTruffleCompilerImpl.Options;
@@ -46,43 +39,21 @@ final class HotSpotTruffleRuntime extends AbstractHotSpotTruffleRuntime {
     }
 
     @Override
-    public <T> T getOptions(Class<T> optionValuesType) {
+    public <T> T getGraalOptions(Class<T> optionValuesType) {
         if (optionValuesType == OptionValues.class) {
             return optionValuesType.cast(HotSpotGraalOptionValues.defaultOptions());
         }
-        return super.getOptions(optionValuesType);
+        return super.getGraalOptions(optionValuesType);
     }
 
     @Override
-    public <T> T convertOptions(Class<T> optionValuesType, Map<String, Object> map) {
-        if (optionValuesType == OptionValues.class) {
-            final EconomicMap<OptionKey<?>, Object> values = OptionValues.newOptionMap();
-            final Iterable<OptionDescriptors> loader = OptionsParser.getOptionsLoader();
-            for (Map.Entry<String, Object> e : map.entrySet()) {
-                final String optionName = e.getKey();
-                final Object optionValue = e.getValue();
-                OptionsParser.parseOption(optionName, optionValue, values, loader);
-            }
-            return optionValuesType.cast(new OptionValues(values));
-        }
-        return super.convertOptions(optionValuesType, map);
-    }
-
-    @Override
-    public Map<String, Object> createInitialOptions() {
-        final UnmodifiableMapCursor<OptionKey<?>, Object> optionValues = getOptions(OptionValues.class).getMap().getEntries();
-        Map<String, Object> res = new HashMap<>();
-        while (optionValues.advance()) {
-            final OptionKey<?> key = optionValues.getKey();
-            Object value = optionValues.getValue();
-            res.put(key.getName(), value);
-        }
-        return res;
+    protected boolean isPrintGraphEnabled() {
+        return DebugOptions.PrintGraph.getValue(getGraalOptions(OptionValues.class)) != DebugOptions.PrintGraphTarget.Disable;
     }
 
     @Override
     protected String initLazyCompilerConfigurationName() {
-        final OptionValues options = getOptions(OptionValues.class);
+        final OptionValues options = getGraalOptions(OptionValues.class);
         String factoryName = Options.TruffleCompilerConfiguration.getValue(options);
         CompilerConfigurationFactory compilerConfigurationFactory = CompilerConfigurationFactory.selectFactory(factoryName, options);
         return compilerConfigurationFactory.getName();